owl-backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.1.0",
]

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-debug-toolbar>=5.2.0",
    "drf-spectacular>=0.28.0",
    "factory-boy>=3.3.3",
    "faker>=37.1.0",
    "isort>=6.0.1",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "pytest-factoryboy>=2.7.0",
]
prod = [
    "gunicorn>=23.0.0",
]
